chore: clean up PEXT implementation
Moved more tests around and fixed up some code.
